Improved data integrity: added exceptions for invalid airport id.

+++ src/main/java/com/selimsahin/amadeus/services/AirportService.java
@@ -2,12 +2,14 @@ package com.selimsahin.amadeus.services;
 
 import com.selimsahin.amadeus.dtos.AirportCreateDto;
 import com.selimsahin.amadeus.dtos.AirportResponseDto;
+import com.selimsahin.amadeus.exceptions.ResourceNotFoundException;
 import com.selimsahin.amadeus.models.Airport;
 import com.selimsahin.amadeus.repositories.AirportRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -24,22 +26,27 @@ public class AirportService {
 
     public Iterable<AirportResponseDto> getAirports() {
         return airportRepository
-            .findAll()
-            .stream()
-            .map(airport -> modelMapper.map(airport, AirportResponseDto.class))
-            .toList();
+                .findAll()
+                .stream()
+                .map(airport -> modelMapper.map(airport, AirportResponseDto.class))
+                .toList();
     }
 
     public AirportResponseDto getAirportById(Long id) {
-        Airport airport = airportRepository.findById(id).orElseThrow(() -> new RuntimeException("Airport not found"));
-        return modelMapper.map(airport, AirportResponseDto.class);
+        Optional<Airport> foundAirport = airportRepository.findById(id);
+
+        if (foundAirport.isEmpty()) {
+            throw new ResourceNotFoundException("Airport not found with id: " + id);
+        }
+
+        return modelMapper.map(foundAirport, AirportResponseDto.class);
     }
 
     public AirportResponseDto updateAirport(Long id, AirportCreateDto airportCreateDto) {
         Optional<Airport> foundAirport = airportRepository.findById(id);
 
         if (foundAirport.isEmpty()) {
-            return null;
+            throw new ResourceNotFoundException("Airport not found with id: " + id);
         }
 
         Airport airport = foundAirport.get();
@@ -50,6 +57,9 @@ public class AirportService {
     }
 
     public void deleteAirport(Long id) {
+        if (!airportRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Airport not found with id: " + id);
+        }
         airportRepository.deleteById(id);
     }
 }
